Fix:Handle outlying use case. TODO: refactor...
Overview:
'editedWord' assumes that cursor is always at the end of a new word.
However, users could add a word directly after an existing one, before retrospectively adding a space between them.
In this case the 'edited' word, identified by the cursor following it, is the existing one.
Therefore, if editedWord has not been transformed, there is a need check if there is a subsequent word and, if so, whether it needs to be transformed.

This outlying use case when editing a document has been functionally handled.

Current code needs to be refactored in to reusable functions for transforming a word and updating compromise / lexical.