Debug: Comments for selected pupil not rendered...
...when clicking on a new name (comments for previous student remain). Comments for previously selected student then render when subsequent student btn is pressed.
Nb. Subjects that the selected student are/aren't assigned to a report group for though are filtered in/out at the right time
- just with incorrect comments.

CAUSE: The correct 'item' was being passed through to PupilSubjectReport, and cld even be updated in the editorState within that module by using a useEffect.
However, the when passed through to the Editor component, the editorState is only used for the initial config of the LexicalComposer. Therefore, the 'new' editorState is not rendered into what are exsiting Editor/LexicalComposer components.

FIX: Make the keys created when mapping through updated SelectedStudentReports unique to the student, thus creating new PupilSubjectReport components > Editor components > LexicalComposer components.
As a result the new PupilSubjectReport components are created and initialised with the 'new comments' for the selected student; this new state is passed through to a new Editor and rendered via the initalConfig for a new LexicalComposer component.